defined _set_bandstack() method
this method is required to transform a handler that hasn't originated
from a bandstack into a bandstacked one. Such a functionality is
required after resamoling of ALL bands into the same resolution and
extent. Otherwise there are problems when adding new bands (meta
information gets lost for the bands to add)

Former-commit-id: 855a5aee578e0a1e5c7c77e2ba3725f07a61d01f